fix: Parent actor index should be used for SBA events
