+++ src-hook/src/hooks/mod.rs
@@ -70,3 +70,47 @@ pub fn actor_type_id(actor_ptr: *const usize) -> u32 {
 pub fn actor_idx(actor_ptr: *const usize) -> u32 {
     unsafe { (actor_ptr.byte_add(0x170) as *const u32).read() }
 }
+
+// Returns the parent entity of the source entity if necessary.
+pub fn get_source_parent(source_type_id: u32, source: *const usize) -> Option<(u32, u32)> {
+    match source_type_id {
+        // Pl0700Ghost -> Pl0700
+        0x2AF678E8 => {
+            let parent_instance = parent_specified_instance_at(source, 0xE48)?;
+
+            Some((actor_type_id(parent_instance), actor_idx(parent_instance)))
+        }
+        // Pl0700GhostSatellite -> Pl0700
+        0x8364C8BC => {
+            let parent_instance = parent_specified_instance_at(source, 0x508)?;
+
+            Some((actor_type_id(parent_instance), actor_idx(parent_instance)))
+        }
+        // Wp1890: Cagliostro's Ouroboros Dragon Sled -> Pl1800
+        0xC9F45042 => {
+            let parent_instance = parent_specified_instance_at(source, 0x578)?;
+            Some((actor_type_id(parent_instance), actor_idx(parent_instance)))
+        }
+        // Pl2000: Id's Dragon Form -> Pl1900
+        0xF5755C0E => {
+            let parent_instance = parent_specified_instance_at(source, 0xD138)?;
+            Some((actor_type_id(parent_instance), actor_idx(parent_instance)))
+        }
+        _ => None,
+    }
+}
+
+// Returns the specified instance of the parent entity.
+// ptr+offset: Entity
+// *(ptr+offset) + 0x70: m_pSpecifiedInstance (Pl0700, Pl1200, etc.)
+fn parent_specified_instance_at(actor_ptr: *const usize, offset: usize) -> Option<*const usize> {
+    unsafe {
+        let info = (actor_ptr.byte_add(offset) as *const *const *const usize).read_unaligned();
+
+        if info == std::ptr::null() {
+            return None;
+        }
+
+        Some(info.byte_add(0x70).read())
+    }
+}
